% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daggle-package.R
\docType{package}
\name{daggle-package}
\alias{daggle}
\alias{daggle-package}
\title{daggle: The daggle app}
\description{
A tool to support learning and teaching the graphical rules of selecting adjustment variables using directed acyclic graphs (DAGs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CBDRH/daggle}
  \item Report bugs at \url{https://github.com/CBDRH/daggle/issues}
}

}
\author{
\strong{Maintainer}: Mark Hanly \email{m.hanly@unsw.edu.au}

}
\keyword{internal}
