% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randDAG.R
\name{randDAG}
\alias{randDAG}
\title{randDAG}
\usage{
randDAG(n, p, seed = 1234)
}
\arguments{
\item{n}{The number of nodes}

\item{p}{The degree of connectivity}

\item{seed}{A random seed}
}
\value{
A `tidy_dagitty` object
}
\description{
Generate a random dag with assigned exposure & outcome and standard naming convention
}
\examples{
dag <- randDAG(6, .5)
}
